\name{fifo_info}
\alias{fifo_info}
\title{Information about the FIFO Page Replacement Algorithm}
\description{
  This function provides detailed information about the FIFO page replacement algorithm, including a description of how it works and the steps involved.
}
\usage{
fifo_info()
}
\arguments{
  None.
}
\value{
  The function does not return any value. It prints the description of the FIFO algorithm and its steps to the console.
}
\details{
  FIFO (First-In-First-Out) is the simplest page replacement algorithm. In FIFO, the page that has been in memory the longest (the oldest) is replaced when a new page needs to be loaded.
  The function outlines the basic steps of the algorithm and provides a brief overview of its history.
}
\examples{
fifo_info()
}
\seealso{
  \code{\link{simulate_fifo}}
}
