\name{lru_info}
\alias{lru_info}
\title{Information about the LRU Page Replacement Algorithm}
\description{
  This function provides detailed information about the LRU page replacement algorithm, including a description of how it works and the steps involved.
}
\usage{
lru_info()
}
\arguments{
  None.
}
\value{
  The function does not return any value. It prints the description of the LRU algorithm and its steps to the console.
}
\details{
  LRU (Least Recently Used) is a page replacement algorithm that replaces the page that has not been used for the longest period of time.
  The function outlines the basic steps of the algorithm and provides a brief overview of its history.
}
\examples{
lru_info()
}
\seealso{
  \code{\link{simulate_lru}}
}
