\name{optimal_info}
\alias{optimal_info}
\title{Information about the Optimal Page Replacement Algorithm}
\description{
  This function provides detailed information about the Optimal page replacement algorithm, including a description of how it works and the steps involved.
}
\usage{
optimal_info()
}
\arguments{
  None.
}
\value{
  The function does not return any value. It prints the description of the Optimal algorithm and its steps to the console.
}
\details{
  The Optimal page replacement algorithm is considered the best among page replacement algorithms because it replaces the page that will not be used for the longest period of time. However, the drawback is that it requires knowledge of the future reference string, which is not feasible in real systems.
}
\examples{
optimal_info()
}
\seealso{
  \code{\link{simulate_optimal}}
}
